Add pixel_mask dtype check

+++ tests/test_file_adapter.py
@@ -15,7 +15,7 @@ DATA_SHAPE_WITH_GAPS_WITH_GEOMETRY = (1100 + 512 + 2, 0 + 1024 + 6)
 IMAGE_SHAPE = DATA_SHAPE_WITH_GAPS_WITH_GEOMETRY
 STACK_IMAGE_SHAPE = (3, *DATA_SHAPE_WITH_GAPS_WITH_GEOMETRY)
 
-pixel_mask_orig = np.random.randint(2, size=DATA_SHAPE, dtype=np.uint32)
+pixel_mask_orig = np.random.randint(2, size=DATA_SHAPE, dtype=np.int64)
 pixel_mask = pixel_mask_orig.astype(bool, copy=True)
 
 image_stack = np.arange(np.prod(STACK_SHAPE), dtype=np.uint16).reshape(STACK_SHAPE[::-1])
